REFACTOR: judgeNew styled components 분리
