+++ src/pages/Judge/JudgeNew/JudgeNewOutModal.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
 import { reset } from '../../../redux/judgeNew';
-import { OutModal, OutModalInner } from '../judge.style';
+import { OutModal, OutModalInner } from './judgeNew.style';
 import xIcon from '../../../assets/img/x-icon.svg';
 
 const JudgeNewOutModal = ({ setIsOut }) => {
