+++ src/pages/Judge/JudgeNew/JudgeSearchResult.jsx
@@ -1,7 +1,7 @@
 /* eslint-disable react/prop-types */
 import React from 'react';
 import InfiniteScroll from 'react-infinite-scroll-component';
-import { JudgeSearchResultContainer } from '../judge.style';
+import { JudgeSearchResultContainer } from './judgeNew.style';
 import defaultIcon from '../../../assets/img/default-icon.svg';
 import { pushRecentSearch } from '../../../libs/utils';
